+++ src/main/java/basball/Application.java
@@ -1,6 +1,6 @@
-package basballRe;
+package basball;
 
-import basballRe.baseball.Running;
+import basball.baseball.Running;
 
 public class Application {
     public static void main(String[] args) {
