+++ src/main/java/basball/baseball/BaseBall.java
@@ -1,6 +1,6 @@
-package basballRe.baseball;
+package basball.baseball;
 
-import basballRe.baseball.model.Balls;
+import basball.baseball.model.Balls;
 
 public class BaseBall {
