+++ src/main/java/basball/baseball/utils/Validation.java
@@ -1,4 +1,4 @@
-package basballRe.baseball.utils;
+package basball.baseball.utils;
 
 import java.util.List;
 import java.util.regex.Pattern;
