+++ src/test/java/basball/view/InputViewTest.java
@@ -1,6 +1,6 @@
-package basballRe.view;
+package basball.view;
 
-import basballRe.baseball.utils.Validation;
+import basball.baseball.utils.Validation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
